NET: Decoupled STMMAC from MDIO interface
The MDIO interface for the STMMAC driver was coupled with the MAC
itself. This did not allow for having a phy which on a MAC's MDIO
bus to be associated with a different MAC for the RGMII interface.
Made the MDIO interface its own platfrom device, which contains the
PHYs. These PHYs can then be registered with the MAC itself.

Signed-off-by: Daniel Vincelette <dvincelette@criticallink.com>