DTS: Created seperate DTS per SOM variant
Created a new inheritance type architecture for the MitySOM 5CSx family.
Now each devkit is drived from a dtsi for that SOM, this should mean
that the dts only has devices that specific to the base board and all
the other devices that are unique only to the SOM are in the SOM's dtsi.