tsc2046: Fixed null pointer exception of interrupt comes in the middle of probe route
There was a race condition in the probe function for the ts2046 touch screen driver
between creating the input device and enabling interrupts. Now the input device is
created before the interrupts are enabled, it was the reverse before which caused
a null pointer exception if a interrupt came in before the input device was created.